deltee works consistently, cleanup/docs
Moves all the seed password stuff into src/seed_password.rs.
seed_password::delete now returns NoEntry if the password is not found.
